Grouped labelled inputs in single file
